Refactor stage structure in detection
