+++ otx/algorithms/detection/configs/detection/mobilenetv2_ssd/semisl/data_pipeline.py
@@ -83,26 +83,6 @@ common_pipeline = [
     ),
 ]
 
-
-img_norm_cfg = dict(mean=[0, 0, 0], std=[255, 255, 255], to_rgb=True)
-train_pipeline = [
-    dict(type="LoadImageFromFile", to_float32=True),
-    dict(type="LoadAnnotations", with_bbox=True),
-    dict(
-        type="PhotoMetricDistortion",
-        brightness_delta=32,
-        contrast_range=(0.5, 1.5),
-        saturation_range=(0.5, 1.5),
-        hue_delta=18,
-    ),
-    dict(type="MinIoURandomCrop", min_ious=(0.1, 0.3, 0.5, 0.7, 0.9), min_crop_size=0.1),
-    dict(type="Resize", img_scale=__img_scale, keep_ratio=False),
-    dict(type="Normalize", **img_norm_cfg),
-    dict(type="RandomFlip", flip_ratio=0.5),
-    dict(type="DefaultFormatBundle"),
-    dict(type="Collect", keys=["img", "gt_bboxes", "gt_labels"]),
-]
-
 train_pipeline = [
     dict(type="LoadImageFromFile"),
     dict(type="LoadAnnotations", with_bbox=True),
