+++ otx/algorithms/detection/tasks/inference.py
@@ -318,11 +318,13 @@ class DetectionInferenceTask(BaseTask, IInferenceTask, IExportTask, IEvaluationT
         return data_cfg
 
     def _update_stage_module(self, stage_module):
-        if self.train_type == TrainType.SEMISUPERVISED:
-            if stage_module == "DetectionTrainer":
-                stage_module = "SemiSLDetectionTrainer"
-            elif stage_module == "DetectionInferrer":
-                stage_module = "SemiSLDetectionInferrer"
+        if stage_module in ["DetectionTrainer", "DetectionInferrer"]:
+            if self.train_type == TrainType.INCREMENTAL:
+                stage_module = "Incr" + stage_module
+            elif self.train_type == TrainType.SEMISUPERVISED:
+                stage_module = "SemiSL" + stage_module
+            else:
+                raise NotImplementedError(f"{self.train_type} do not have own trainer or inferrer")
         return stage_module
 
     def _add_predictions_to_dataset(self, prediction_results, dataset, confidence_threshold=0.0):
