+++ otx/mpa/det/incremental/__init__.py
@@ -2,6 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
+from .inferrer import IncrDetectionInferrer
 from .stage import IncrDetectionStage
+from .trainer import IncrDetectionTrainer
 
-__all__ = ["IncrDetectionStage"]
+__all__ = ["IncrDetectionStage", "IncrDetectionTrainer", "IncrDetectionInferrer"]
