+++ otx/mpa/det/incremental/trainer.py
@@ -3,22 +3,18 @@
 #
 
 import glob
-import numbers
-import os
 import os.path as osp
 import time
 
-import torch
-import torch.distributed as dist
-import torch.multiprocessing as mp
 from mmcv.utils import get_git_hash
 from mmdet import __version__
 from mmdet.apis import train_detector
 from mmdet.datasets import build_dataset
 from mmdet.models import build_detector
 from mmdet.utils import collect_env
+from torch import nn
 
-from otx.mpa.det.semisl import SemiSLDetectionStage
+from otx.mpa.det.incremental import IncrDetectionStage
 from otx.mpa.det.trainer import DetectionTrainer
 from otx.mpa.modules.utils.task_adapt import extract_anchor_ratio
 from otx.mpa.registry import STAGES
@@ -31,7 +27,8 @@ from otx.mpa.utils.logger import get_logger
 logger = get_logger()
 
 
+# FIXME DetectionTrainer does not inherit from stage
 @STAGES.register_module()
-class SemiSLDetectionTrainer(SemiSLDetectionStage, DetectionTrainer):
+class IncrDetectionTrainer(IncrDetectionStage, DetectionTrainer):
     def __init__(self, **kwargs):
-        SemiSLDetectionStage.__init__(self, **kwargs)
+        IncrDetectionStage.__init__(self, **kwargs)
